feature(order security) add protection for order security

+++ server/api/orders.js
@@ -9,7 +9,7 @@ router.use('/line-items', require('./line-items'))
 // this route gets all orders. It's accessible to only ADMIN users. 
 router.get('/', async (req, res, next) => {
   try {
-      //THIS NEEDS TO BE PROTECTS FOR ADMINS ONLY
+      //THIS NEEDS TO BE PROTECTED FOR ADMINS ONLY
       res.json(await Order.findAll({
         include: [{model: LineItem}]
       }))
@@ -22,15 +22,22 @@ router.get('/', async (req, res, next) => {
 router.get('/:userId', async (req, res, next) => {
   try {
     if(!isNaN(req.params.userId)) { //number means it's a user id
-      res.json(await Order.findOne({
-        where: {
-          userId: req.params.userId
-        },
-        include: [{model: LineItem, include: [
-          {model: Item}
-        ]}]
-      }))
+      if(req.user && req.user.id != req.params.userId){
+        //user is getting a 401 because they do not have the same id as the id for whom the order belongs to. Therefor they should not have access to this data.
+        res.status(401).send('These bugs are not your bugs!', )
+      } else {
+        //user has the same id as the id for whom the order belongs to. Therefor they should not have access to this data.
+        res.json(await Order.findOne({
+          where: {
+            userId: req.params.userId
+          },
+          include: [{model: LineItem, include: [
+            {model: Item}
+          ]}]
+        }))
+      }
     } else { // if it's not a number, it's a string. It's a guest user
+      //I DO NOT KNOW HOW TO HIT THESE...
       res.json(await Order.findOne({
         where: {
           guestSessionId: req.params.userId
@@ -68,11 +75,17 @@ router.post('/', async (req, res, next) => {
 //this would be used if a user empties their cart.
 router.delete('/:orderId', async (req, res, next) => {
   try {
-    await Order.destroy({
-      where: {
-        id: req.params.orderId
-      }
-    })
+    if(req.user.id != req.params.userId){
+      //user is getting a 401 because they do not have the same id as the id for whom the order belongs to. Therefor they should not be able to delete this order.
+      res.sendStatus(401)
+    } else {
+      //user has the same id as the id for whom the order belongs to. Therefor they should not have access to delete this data.
+      await Order.destroy({
+        where: {
+          id: req.params.orderId
+        }
+      })
+    }
     res.send('order has been deleted from the database')
   } catch (err) {
     next(err);
@@ -85,16 +98,22 @@ router.delete('/:orderId', async (req, res, next) => {
 router.put('/:orderId', async (req, res, next) => {
   try {
     if(!isNaN(req.params.orderId)) { //number means it's a user id
-      await Order.update(
-        req.body,
-        {where: {
-          id: req.params.orderId  
-      }})
-      res.json(await Order.findById(req.params.orderId, {
-        include: [{model: LineItem, include: [
-          {model: Item}
-        ]}]
-      }))
+      if(req.user.id != req.params.userId){
+        //user is getting a 401 because they do not have the same id as the id for whom the order belongs to. Therefor they should not be able to make posts to this order.
+        res.sendStatus(401)
+      } else {
+        //user has the same id as the id for whom the order belongs to. Therefor they should not have access to this data.
+        await Order.update(
+          req.body,
+          {where: {
+            id: req.params.orderId  
+        }})
+        res.json(await Order.findById(req.params.orderId, {
+          include: [{model: LineItem, include: [
+            {model: Item}
+          ]}]
+        }))
+      }
     } else { // if it's not a number, it's a string. It's a guest user
       await Order.update(
         req.body,
